% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoviceDeveloperResources.R
\name{conflictOfInterest}
\alias{conflictOfInterest}
\title{conflictOfInterest}
\usage{
conflictOfInterest()
}
\value{
returns a list identifying the conflicts for each conflicted function
}
\description{
determine if there are any conflicts
between the functions in the attached packages and the R global environment
}
\examples{
l<-conflictOfInterest()

}
