% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoviceDeveloperResources.R
\name{conflictOfInterestRestricted}
\alias{conflictOfInterestRestricted}
\title{conflictOfInterestRestricted}
\usage{
conflictOfInterestRestricted(packs)
}
\arguments{
\item{packs}{character vector of user-specified packages}
}
\value{
returns a subset of the return value of conflictOfInterest()
}
\description{
restrict the conflicted functions (retrieved by conflictOfInterest())
to those in user-specified packages
}
