% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoviceDeveloperResources.R
\name{dateTable}
\alias{dateTable}
\title{dateTable}
\usage{
dateTable(dt, f, when)
}
\arguments{
\item{dt}{date table in list format}

\item{f}{character string containing the full path name of the file}

\item{when}{character string either "before" or "after"}
}
\value{
updated version of dt
}
\description{
list of file dates
}
\details{
allows the user to confirm that the .tar.gz and the library packages are not left-overs
}
\examples{
\dontrun{
# you need to specify dir, f that is on your own computer !!
dt<-list()
dir<-"~/personal/hearts/hearts_card_game_bayesian_inference/packages"
f<-sprintf("\%s/\%s",dir,"NoviceDeveloperResources_1.1.0.tar.gz")
when<-"before"
dateTable(dt,f,when)
}

}
