% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoviceDeveloperResources.R
\name{fExistsAge}
\alias{fExistsAge}
\title{fExistsAge}
\usage{
fExistsAge(f)
}
\arguments{
\item{f}{character string containing the path name of the file}
}
\value{
returns no values,but has side effect of terminating if file is not valid
}
\description{
check for newly created file existence and age
}
\examples{
\dontrun{
# you need to specify f that is on your own computer !!
dir1<-"/Users/barryzeeberg/personal/hearts"
dir2<-"hearts_card_game_bayesian_inference/packages"
f<-"NoviceDeveloperResources_1.1.0.tar.gz"
fExistsAge(sprintf("\%s/\%s/\%s",dir1,dir2,f))
}

}
