% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoviceDeveloperResources.R
\name{formatSearchPack}
\alias{formatSearchPack}
\title{formatSearchPack}
\usage{
formatSearchPack(pack)
}
\arguments{
\item{pack}{list of character strings containing the names of a package}
}
\value{
returns list of package names formatted it as listed in the search path
}
\description{
given a package name, reformat it as listed in the search path
}
\examples{
pack<-c("retrieve","tcpflow")
formatSearchPack(pack)

}
