% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoviceDeveloperResources.R
\name{getSysLib}
\alias{getSysLib}
\title{getSysLib}
\usage{
getSysLib()
}
\value{
returns a character string containing the path name for the designated library
}
\description{
For consistency, make sure that we are always using the designated library rather than a random library
}
\examples{
if(interactive()) {
  sysLib<-getSysLib()
}

}
