% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoviceDeveloperResources.R
\name{inSearchPath}
\alias{inSearchPath}
\title{inSearchPack}
\usage{
inSearchPath(pack)
}
\arguments{
\item{pack}{list of character strings containing the name of a package}
}
\value{
returns list of Booleans TRUE if the package is listed in the search path
}
\description{
is the package listed in the search path
}
\examples{
\dontrun{
# you need to specify packs that are on your own computer !!
pack<-c("retrieve","tcpflow")
inSearchPath(pack)
}

}
