% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoviceDeveloperResources.R
\name{zload}
\alias{zload}
\title{zload}
\usage{
zload(lib, pack, dt)
}
\arguments{
\item{lib}{character string containing the name of the user-designated library}

\item{pack}{character string containing the name of a package}

\item{dt}{date table in list format}
}
\value{
returns updated version of date table
}
\description{
detach old version of package from search path, load new version, and validate
}
\examples{
\dontrun{
# you need to specify packs that are on your own computer !!
lib<-getSysLib()
dt<-list()
zload(lib,"NoviceDeveloperResources",dt)
}

}
