% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoviceDeveloperResources.R
\name{checkBuildInstallSourcePackage}
\alias{checkBuildInstallSourcePackage}
\title{checkBuildInstallSourcePackage}
\usage{
checkBuildInstallSourcePackage(dir, packs)
}
\arguments{
\item{dir}{character string containing the path name of the directory holding the package folders}

\item{packs}{character vector of the names of the packages}
}
\value{
returns a character string that can be printed using "echo" and then
copy and paste by the user to load or update the packages
}
\description{
check, build, and install local source package
}
\details{
I wanted to include "library(packs)" in the program, but this is not allowed.
The user can "cat" the return value and copy and paste to facilitate doing
this manually.
}
